Various small fixes towards building with other compilers
* Fix various non-GCC warnings (no effect on GCC)
* Reduce use of typeof since non GCC compilers may not support it
* Introduce PICO_C_COMPILER_IS_GNU, PICO_C_COMPILER_IS_CLANG, PICO_C_COMPILER_IS_IAR to CMake as if (CMAKE_C_COMPILER_ID STREQUAL "xxx") is a bit verbose
* Use "unified_asm" macro for all inline asm (it is "volatile __asm" on GNU with a .syntex unified)
* Use NOLOAD instead of COPY in linker scripts (arguably more correct anyway)
* Use the same style for setting _etext in all 4 linker scripts (to the beginning of .data). Clang aligns .data on a 16 byte boundary. Note ideally we'd
  add a new symbol __data_source, however that would break backwards compatibility with existing user linker scripts
* Use "a" for .stack, .heap sections because clang complains otherwise, and they are explicitly NOLOAD anyway
* Avoid duplicating __sev, __wfe, __wfi which Clang sometimes seems to provide as built-ins
* Add missing kitchen_sink_blocked_ram binary
* Allow build with LLVM Embedded Toolchain Form ARM v 14.0.0 (unsupported atm)

+++ test/pico_divider_test/pico_divider_nesting_test.c
@@ -34,16 +34,16 @@ bool timer_callback(repeating_timer_t *t) {
             FAILED();
         }
     }
-    float fz = z;
+    float fz = (float)z;
     float fa = fz / 11.0f;
     float fb = fmodf(fz, 11.0f);
-    if (fabsf(fz - (fa * 11.0 + fb) > 1e-9)) {
+    if (fabs(fz - (fa * 11.0 + fb)) > 1e-9) {
         FAILED();
     }
     double dz = z;
     double da = dz / 11.0;
     double db = fmod(dz, 11.0);
-    if (fabsf(dz - (da * 11.0 + db) > 1e-9)) {
+    if (fabs(dz - (da * 11.0 + db)) > 1e-9) {
         FAILED();
     }
